requirements.txt
#pip install -r requirements.txt
sqlalchemy==1.4.37
pymysql==1.0.2
pandas==1.4.2
python-dotenv==0.20.0
psycopg2==2.9.3
